Refactor: changing a file name since it breaks with code style
