+++ common/CMakeLists.txt
@@ -16,7 +16,7 @@ project (mmotd_common)
 
 add_library (
     ${MMOTD_TARGET_NAME} STATIC
-    src/chrono-io.cpp
+    src/chrono_io.cpp
     src/human_size.cpp
     src/platform_error.cpp
     )
